Move PSBT stuff to its own module
